requirements.txt
aiogram>=3.13.0,<4
motor>=3.6.0
aiohttp>=3.9.0
matplotlib>=3.8.0
pymongo>=4.8.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
mongomock>=4.1.0
